Dataset: Aristolochia

The data come from an experiment conducted at the Seed Analysis Laboratory of the Agricultural Sciences Center of the State University of Londrina, in which five temperatures (15, 20, 25, 30 and 35C) were evaluated in the germination of Aristolochia elegans. The experiment was conducted in a completely randomized design with four replications of 25 seeds each.

data("aristolochia")

Format

data.frame containing data set

trat

Numeric vector with temperature

resp

Numeric vector with response

Author

Hugo Roldi Guariz

Examples

data(aristolochia)